+++ R/upper_tier_region_ctry_lookup.R
@@ -1,0 +1,12 @@
+#' A lookup of upper-tier Local Authorities (EW) to their Region (if any) and Country
+#'
+#' @format A data frame with 174 rows and 6 variables:
+#' \describe{
+#'   \item{utla21cd}{UTLA code}
+#'   \item{utla21nm}{UTLA name}
+#'   \item{rgn21cd}{Region code}
+#'   \item{rgn21nm}{Region name}
+#'   \item{ctry21cd}{Country code}
+#'   \item{ctry21nm}{Country name}
+#' }
+"upper_tier_region_ctry_lookup"